week-2/requirements.txt
# ── Week 1 (carried forward) ──────────────────────
pymupdf==1.25.3
chromadb==1.0.15
sentence-transformers==3.0.1
openai==1.95.1
python-dotenv==1.0.1
tqdm==4.67.1
numpy==1.26.4

# ── Week 2 (NEW) ─────────────────────────────────
anthropic>=0.45.0
mistralai>=1.5.0
cohere>=5.13.0
